+++ src/lib/models/reflections/signature.ts
@@ -4,9 +4,11 @@ import type { ParameterReflection } from "./parameter";
 import type { TypeParameterReflection } from "./type-parameter";
 import type { DeclarationReflection } from "./declaration";
 import type { ReflectionKind } from "./kind";
-import type { Serializer, JSONOutput } from "../../serialization";
+import type { Serializer, JSONOutput, Deserializer } from "../../serialization";
 
 export class SignatureReflection extends Reflection {
+    readonly variant = "signature";
+
     constructor(
         name: string,
         kind: SignatureReflection["kind"],
@@ -109,6 +111,7 @@ export class SignatureReflection extends Reflection {
     override toObject(serializer: Serializer): JSONOutput.SignatureReflection {
         return {
             ...super.toObject(serializer),
+            variant: this.variant,
             typeParameter: serializer.toObjectsOptional(this.typeParameters),
             parameters: serializer.toObjectsOptional(this.parameters),
             type: serializer.toObject(this.type),
@@ -116,5 +119,23 @@ export class SignatureReflection extends Reflection {
             inheritedFrom: serializer.toObject(this.inheritedFrom),
             implementationOf: serializer.toObject(this.implementationOf),
         };
+    }
+
+    override fromObject(
+        de: Deserializer,
+        obj: JSONOutput.SignatureReflection
+    ): void {
+        super.fromObject(de, obj);
+
+        this.typeParameters = de.reviveMany(obj.typeParameter, (t) =>
+            de.constructReflection(t)
+        );
+        this.parameters = de.reviveMany(obj.parameters, (t) =>
+            de.constructReflection(t)
+        );
+        this.type = de.reviveType(obj.type);
+        this.overwrites = de.reviveType(obj.overwrites);
+        this.inheritedFrom = de.reviveType(obj.inheritedFrom);
+        this.implementationOf = de.reviveType(obj.implementationOf);
     }
 }