+++ src/lib/serialization/index.ts
@@ -1,7 +1,10 @@
 export type { SerializerComponent } from "./components";
+export {
+    Deserializer,
+    type Deserializable,
+    type DeserializerComponent,
+} from "./deserializer";
 export { SerializeEvent } from "./events";
 export type { SerializeEventData } from "./events";
+export * as JSONOutput from "./schema";
 export { Serializer } from "./serializer";
-export { JSONOutput };
-
-import * as JSONOutput from "./schema";