+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 -   Removed special case for `--plugin none`, to disable plugin auto discovery, set `"plugin": []` in a config file.
 -   `Application.bootstrap` is now async to support ESM based plugins, #1635.
 -   Removed internal `ReferenceType.getSymbol`, reference types no longer reference the `ts.Symbol` to enable generation from serialized JSON.
+-   `OptionsReader.priority` has been renamed to `OptionsReader.order` to more accurately reflect how it works.
+-   `id: -1` will not be included in the JSON output for reference types which are intentionally broken.
 
 ### Features